Extract text generation into dedicated server module
- Move git text-generation layers and tests under `src/textGeneration`
- Route server and provider drivers to the new module
- Replace the old git-scoped `TextGenerationLive` registry layer
