+++ apps/server/src/textGeneration/TextGeneration.ts
@@ -1,18 +1,13 @@
-/**
- * TextGeneration - Effect service contract for AI-generated Git content.
- *
- * Generates commit messages and pull request titles/bodies from repository
- * context prepared by Git services.
- *
- * @module TextGeneration
- */
-import { Context } from "effect";
-import type { Effect } from "effect";
-import type { ChatAttachment, ModelSelection } from "@t3tools/contracts";
+import { Context, Effect, Layer } from "effect";
+import type { ChatAttachment, ModelSelection, ProviderInstanceId } from "@t3tools/contracts";
+import { TextGenerationError } from "@t3tools/contracts";
 
-import type { TextGenerationError } from "@t3tools/contracts";
+import {
+  ProviderInstanceRegistry,
+  type ProviderInstanceRegistryShape,
+} from "../provider/Services/ProviderInstanceRegistry.ts";
+import type { ProviderInstance } from "../provider/ProviderDriver.ts";
 
-/** Providers that support git text generation (commit messages, PR content, branch names). */
 export type TextGenerationProvider = "codex" | "claudeAgent" | "cursor" | "opencode";
 
 export interface CommitMessageGenerationInput {
@@ -119,5 +114,58 @@ export interface TextGenerationShape {
  * TextGeneration - Service tag for commit and PR text generation.
  */
 export class TextGeneration extends Context.Service<TextGeneration, TextGenerationShape>()(
-  "t3/git/Services/TextGeneration",
+  "t3/text-generation/TextGeneration",
 ) {}
+
+type TextGenerationOp =
+  | "generateCommitMessage"
+  | "generatePrContent"
+  | "generateBranchName"
+  | "generateThreadTitle";
+
+const resolveInstance = (
+  registry: ProviderInstanceRegistryShape,
+  operation: TextGenerationOp,
+  instanceId: ProviderInstanceId,
+): Effect.Effect<ProviderInstance["textGeneration"], TextGenerationError> =>
+  registry.getInstance(instanceId).pipe(
+    Effect.flatMap((instance) =>
+      instance
+        ? Effect.succeed(instance.textGeneration)
+        : Effect.fail(
+            new TextGenerationError({
+              operation,
+              detail: `No provider instance registered for id '${instanceId}'.`,
+            }),
+          ),
+    ),
+  );
+
+export const makeTextGenerationFromRegistry = (
+  registry: ProviderInstanceRegistryShape,
+): TextGenerationShape => ({
+  generateCommitMessage: (input) =>
+    resolveInstance(registry, "generateCommitMessage", input.modelSelection.instanceId).pipe(
+      Effect.flatMap((textGeneration) => textGeneration.generateCommitMessage(input)),
+    ),
+  generatePrContent: (input) =>
+    resolveInstance(registry, "generatePrContent", input.modelSelection.instanceId).pipe(
+      Effect.flatMap((textGeneration) => textGeneration.generatePrContent(input)),
+    ),
+  generateBranchName: (input) =>
+    resolveInstance(registry, "generateBranchName", input.modelSelection.instanceId).pipe(
+      Effect.flatMap((textGeneration) => textGeneration.generateBranchName(input)),
+    ),
+  generateThreadTitle: (input) =>
+    resolveInstance(registry, "generateThreadTitle", input.modelSelection.instanceId).pipe(
+      Effect.flatMap((textGeneration) => textGeneration.generateThreadTitle(input)),
+    ),
+});
+
+export const layer = Layer.effect(
+  TextGeneration,
+  Effect.gen(function* () {
+    const registry = yield* ProviderInstanceRegistry;
+    return makeTextGenerationFromRegistry(registry);
+  }),
+);
